UHF-8525: Fixed annotations, moved test to correct namespace, added test for CacheTagInvalidator service
