+++ tests/src/Unit/Cache/CacheInvalidatorSubscriberTest.php
@@ -4,9 +4,8 @@ declare(strict_types = 1);
 
 namespace Drupal\Tests\helfi_api_base\Unit\Cache;
 
-use Drupal\Core\Cache\CacheTagsInvalidatorInterface;
-use Drupal\helfi_api_base\Azure\PubSub\PubSubMessage;
-use Drupal\helfi_api_base\EventSubscriber\CacheInvalidatorSubscriber;
+use Drupal\helfi_api_base\Azure\PubSub\PubSubManagerInterface;
+use Drupal\helfi_api_base\Cache\CacheTagInvalidator;
 use Drupal\Tests\UnitTestCase;
 use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
@@ -24,21 +23,11 @@ class CacheInvalidatorSubscriberTest extends UnitTestCase {
    * @covers ::invalidateTags
    */
   public function testInvalidCacheTags() : void {
-    $invalidator = $this->prophesize(CacheTagsInvalidatorInterface::class);
-    $invalidator->invalidateTags(Argument::any())->shouldNotBeCalled();
-    $sut = new CacheInvalidatorSubscriber($invalidator->reveal());
-    $sut->onReceive(new PubSubMessage([]));
-  }
+    $client = $this->prophesize(PubSubManagerInterface::class);
+    $client->sendMessage(Argument::any())->shouldBeCalledTimes(1);
 
-  /**
-   * @covers ::__construct
-   * @covers ::invalidateTags
-   */
-  public function testValidCacheTags() : void {
-    $invalidator = $this->prophesize(CacheTagsInvalidatorInterface::class);
-    $invalidator->invalidateTags(['node:123'])->shouldBeCalled();
-    $sut = new CacheInvalidatorSubscriber($invalidator->reveal());
-    $sut->onReceive(new PubSubMessage(['tags' => ['node:123']]));
+    $sut = new CacheTagInvalidator($client->reveal());
+    $sut->invalidateTags(['node:123']);
   }
 
 }
